fix swag by put commants outside function
