+++ .github/workflows/ci.yml
@@ -44,7 +44,7 @@ jobs:
       env:
         GOPATH: /home/runner/go
     - name: Generate Docs
-      run: /home/runner/go/bin/swag init --parseDependency --parseDepth 2 --parseInternal -g web/main.go
+      run: /home/runner/go/bin/swag init --parseDependency --parseDepth 2 --parseInternal -g web/api/docs.go -o web/docs
     - name: Test
       run: go test $(go list ./... | grep -v /vendor/) -v -failfast -p 1 -coverprofile .testCoverage.txt
     - name: Show Coverage
@@ -59,7 +59,7 @@ jobs:
       env:
         GOPATH: /home/runner/go
     - name: Generate Docs
-      run: /home/runner/go/bin/swag init --parseDependency --parseDepth 2 --parseInternal -g web/main.go
+      run: /home/runner/go/bin/swag init --parseDependency --parseDepth 2 --parseInternal -g web/api/docs.go -o web/docs
     - name: Build
       run: go build -v
     - name: Archive build artifacts
