+++ web/api/docs.go
@@ -1,0 +1,13 @@
+package api
+
+// Bind all routing
+// @title Mediathek API
+// @version 1.0
+// @description This is the first version of the OpenSource Mediathek of Tomorrow
+// @termsOfService http://swagger.io/terms/
+// -host v2.media.kukoon.de
+// @BasePath /
+//
+// @securityDefinitions.apikey ApiKeyAuth
+// @in header
+// @name Authorization
