+++ web/docs/main.go
@@ -5,7 +5,6 @@ import (
 	swaggerFiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
 
-	_ "github.com/Kukoon/media-server/docs"
 	"github.com/Kukoon/media-server/web"
 )
